Create gamelogic.js and export currentBoard to store.js
Take the game logic out of app.js and move it to a new file, called
gamelogic.js. Fix all the requires so that they work correctly. Use
store.js to keep track of the currentBoard variable so that I don't have
to pass it around from file to file.